[go-gui] Removed weird version specifier in Java dependency
I'm actually not sure what this thing was even *trying* to accomplish;
normally this is used for setting a scope, not a version. It was failing
for me on a machine with the latest `brew cask install java` before, and
now it's fine.

In any case, it's certainly *NOT* true that go-gui requires Java 1.6
specifically to run. Any later versions (like the latest 1.8) work just
fine, and you'd be hard-pressed to find a Mac with Java <= 1.5 anyway.